# frontend/openapi.json
{
  "components": {
    "schemas": {
      "AgentTraceEntry": {
        "additionalProperties": false,
        "properties": {
          "agent": {
            "title": "Agent",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": ["agent", "status", "summary"],
        "title": "AgentTraceEntry",
        "type": "object"
      },
      "AnalysisResponse": {
        "additionalProperties": false,
        "properties": {
          "agent_trace": {
            "items": {
              "$ref": "#/components/schemas/AgentTraceEntry"
            },
            "title": "Agent Trace",
            "type": "array"
          },
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/EvidenceArtifact"
            },
            "title": "Artifacts",
            "type": "array"
          },
          "controls": {
            "items": {
              "$ref": "#/components/schemas/ControlReportResult"
            },
            "title": "Controls",
            "type": "array"
          },
          "disclaimer": {
            "title": "Disclaimer",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "summary": {
            "$ref": "#/components/schemas/AnalysisSummary"
          },
          "uploaded_filename": {
            "title": "Uploaded Filename",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "uploaded_filename",
          "summary",
          "artifacts",
          "controls",
          "agent_trace",
          "disclaimer"
        ],
        "title": "AnalysisResponse",
        "type": "object"
      },
      "AnalysisSummary": {
        "additionalProperties": false,
        "properties": {
          "needs_prep_count": {
            "title": "Needs Prep Count",
            "type": "integer"
          },
          "partial_count": {
            "title": "Partial Count",
            "type": "integer"
          },
          "ready_count": {
            "title": "Ready Count",
            "type": "integer"
          },
          "top_auditor_questions": {
            "items": {
              "type": "string"
            },
            "title": "Top Auditor Questions",
            "type": "array"
          },
          "total_controls": {
            "title": "Total Controls",
            "type": "integer"
          },
          "total_gap_count": {
            "title": "Total Gap Count",
            "type": "integer"
          }
        },
        "required": [
          "total_controls",
          "ready_count",
          "partial_count",
          "needs_prep_count",
          "total_gap_count",
          "top_auditor_questions"
        ],
        "title": "AnalysisSummary",
        "type": "object"
      },
      "ApiErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": ["code", "message"],
        "title": "ApiErrorDetail",
        "type": "object"
      },
      "Body_analyzeEvidence": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": ["file"],
        "title": "Body_analyzeEvidence",
        "type": "object"
      },
      "Body_streamAnalyzeEvidence": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": ["file"],
        "title": "Body_streamAnalyzeEvidence",
        "type": "object"
      },
      "ComplianceScores": {
        "additionalProperties": false,
        "properties": {
          "compliance_effort": {
            "default": 0,
            "title": "Compliance Effort",
            "type": "integer"
          },
          "letter_of_law": {
            "default": 0,
            "title": "Letter Of Law",
            "type": "integer"
          },
          "spirit_of_law": {
            "default": 0,
            "title": "Spirit Of Law",
            "type": "integer"
          }
        },
        "title": "ComplianceScores",
        "type": "object"
      },
...
      "ControlEvidenceFound": {
        "additionalProperties": false,
        "properties": {
          "artifact": {
            "title": "Artifact",
            "type": "string"
          },
          "claim": {
            "title": "Claim",
            "type": "string"
          },
          "hash": {
            "title": "Hash",
            "type": "string"
          },
          "location": {
            "title": "Location",
            "type": "string"
          }
        },
        "required": ["claim", "artifact", "location", "hash"],
        "title": "ControlEvidenceFound",
        "type": "object"
      },
      "ControlReportResult": {
        "additionalProperties": false,
        "properties": {
          "agent_plan": {
            "items": {
              "type": "string"
            },
            "title": "Agent Plan",
            "type": "array"
          },
          "artifact": {
            "title": "Artifact",
            "type": "string"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "confidence_rationale": {
            "title": "Confidence Rationale",
            "type": "string"
          },
          "evidence_found": {
            "items": {
              "$ref": "#/components/schemas/ControlEvidenceFound"
            },
            "title": "Evidence Found",
            "type": "array"
          },
          "gaps": {
            "items": {
              "type": "string"
            },
            "title": "Gaps",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "llm_reasoning": {
            "default": "",
            "title": "Llm Reasoning",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "priority": {
            "title": "Priority",
            "type": "string"
          },
          "provenance": {
            "title": "Provenance",
            "type": "string"
          },
          "reasoning": {
            "items": {
              "type": "string"
            },
            "title": "Reasoning",
            "type": "array"
          },
          "regulation_story": {
            "title": "Regulation Story",
            "type": "string"
          },
          "regulatory_citations": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Regulatory Citations",
            "type": "array"
          },
          "reviewer_question": {
            "title": "Reviewer Question",
            "type": "string"
          },
          "risk_summary": {
            "title": "Risk Summary",
            "type": "string"
          },
          "scores": {
            "$ref": "#/components/schemas/ComplianceScores",
            "default": {
              "compliance_effort": 0,
              "letter_of_law": 0,
              "spirit_of_law": 0
            }
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "suggestion": {
            "title": "Suggestion",
            "type": "string"
          },
          "tool_trace": {
            "items": {
              "$ref": "#/components/schemas/ToolTraceEntry"
            },
            "title": "Tool Trace",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "priority",
          "status",
          "confidence",
          "regulation_story",
          "artifact",
          "reasoning",
          "reviewer_question",
          "suggestion",
          "provenance",
          "risk_summary",
          "evidence_found",
          "gaps",
          "agent_plan",
          "tool_trace",
          "confidence_rationale"
        ],
        "title": "ControlReportResult",
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorDetail"
          }
        },
        "required": ["error"],
        "title": "ErrorResponse",
        "type": "object"
      },
      "EvidenceArtifact": {
        "additionalProperties": false,
        "properties": {
          "chunk_count": {
            "title": "Chunk Count",
            "type": "integer"
          },
          "extension": {
            "title": "Extension",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": ["name", "extension", "chunk_count"],
        "title": "EvidenceArtifact",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Requirement": {
        "additionalProperties": false,
        "properties": {
          "audit_frequency": {
            "default": "",
            "title": "Audit Frequency",
            "type": "string"
          },
          "compliance_mapping_raw": {
            "default": "",
            "title": "Compliance Mapping Raw",
            "type": "string"
          },
          "evidence_source": {
            "default": "",
            "title": "Evidence Source",
            "type": "string"
          },
          "frameworks": {
            "items": {
              "type": "string"
            },
            "title": "Frameworks",
            "type": "array"
          },
          "freshness_sla_days": {
            "default": 90,
            "title": "Freshness Sla Days",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "policy_id": {
            "title": "Policy Id",
            "type": "string"
          },
          "policy_name": {
            "title": "Policy Name",
            "type": "string"
          },
          "raw_text": {
            "title": "Raw Text",
            "type": "string"
          },
          "responsible": {
            "default": "",
            "title": "Responsible",
            "type": "string"
          },
          "scope": {
            "default": "",
            "title": "Scope",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": ["id", "policy_id", "policy_name", "text", "raw_text"],
        "title": "Requirement",
        "type": "object"
      },
      "RequirementCatalogResponse": {
        "additionalProperties": false,
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "requirements": {
            "items": {
              "$ref": "#/components/schemas/Requirement"
            },
            "title": "Requirements",
            "type": "array"
          },
          "source": {
            "title": "Source",
            "type": "string"
          }
        },
        "required": ["source", "count"],
        "title": "RequirementCatalogResponse",
        "type": "object"
      },
...
      "ToolTraceEntry": {
        "additionalProperties": false,
        "properties": {
          "input": {
            "title": "Input",
            "type": "string"
          },
          "result": {
            "title": "Result",
            "type": "string"
          },
          "tool": {
            "title": "Tool",
            "type": "string"
          }
        },
        "required": ["tool", "input", "result"],
        "title": "ToolTraceEntry",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Compliance readiness analysis backend for CompAud.",
    "title": "CompAud Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/analyze": {
      "post": {
        "operationId": "analyzeEvidence",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyzeEvidence"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid evidence upload."
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unsupported evidence file type."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyze uploaded compliance evidence",
        "tags": ["analysis"]
      }
    },
    "/analyze/stream": {
      "post": {
        "operationId": "streamAnalyzeEvidence",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_streamAnalyzeEvidence"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream uploaded compliance evidence analysis progress",
        "tags": ["analysis"]
      }
    },
...
    "/knowledge/context/{control_id}": {
      "get": {
        "operationId": "getControlRegulatoryContext",
        "parameters": [
          {
            "description": "Control ID, e.g. FT-IAM-01",
            "in": "path",
            "name": "control_id",
            "required": true,
            "schema": {
              "description": "Control ID, e.g. FT-IAM-01",
              "title": "Control Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Getcontrolregulatorycontext",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get full regulatory context for a control",
        "tags": ["knowledge"]
      }
    },
    "/knowledge/priority/{priority}": {
      "get": {
        "operationId": "getRulesByPriority",
        "parameters": [
          {
            "description": "Priority level, e.g. P1_Critical",
            "in": "path",
            "name": "priority",
            "required": true,
            "schema": {
              "description": "Priority level, e.g. P1_Critical",
              "title": "Priority",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Getrulesbypriority",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get all rules at a given priority level",
        "tags": ["knowledge"]
      }
    },
    "/knowledge/rules/{control_id}": {
      "get": {
        "operationId": "getControlGoverningRules",
        "parameters": [
          {
            "description": "Control ID, e.g. FT-IAM-01",
            "in": "path",
            "name": "control_id",
            "required": true,
            "schema": {
              "description": "Control ID, e.g. FT-IAM-01",
              "title": "Control Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Getcontrolgoverningrules",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get governing rules for a control from official regulatory sources",
        "tags": ["knowledge"]
      }
    },
    "/knowledge/search": {
      "get": {
        "operationId": "searchRules",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Searchrules",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Full-text search across regulatory rule text",
        "tags": ["knowledge"]
      }
    },
...
    "/ps3/requirements/{requirement_id}": {
      "get": {
        "operationId": "getPs3Requirement",
        "parameters": [
          {
            "in": "path",
            "name": "requirement_id",
            "required": true,
            "schema": {
              "title": "Requirement Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementCatalogResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get a single parsed requirement by minted id",
        "tags": ["ps3"]
      }
    }
  },
...
}